src/test/java/application/services/UserServiceTest.java
package application.services;

import event.recommendation.system.entities.user.User;
import event.recommendation.system.repositories.UserRepository;
import event.recommendation.system.services.MailSenderService;
import event.recommendation.system.services.user.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import static org.junit.jupiter.api.Assertions.*;

@ExtendWith(SpringExtension.class)
@SpringBootTest
class UserServiceTest {

    @Mock
    private UserRepository userRepository;
    @Mock
    private MailSenderService mailSenderService;
    @Mock
    private PasswordEncoder passwordEncoder;
    @InjectMocks
    private UserService userService;
    private User user;

    @BeforeEach
    void createUser(){
        this.user = new User();
    }

    @Test
    void userServiceIsNotNull(){
        assertNotNull(userService);
    }

    @Test
    void notEmptyEmailTest(){
        user.setEmail("non empty");
        assertFalse(userService.isUserEmailEmpty(user));
    }

    @Test
    void emptyEmailTest(){
        user.setEmail("");
        assertTrue(userService.isUserEmailEmpty(user));
    }

    @Test
    public void isUserExist(){
        user.setEmail("test.test@test.test");

        userRepository.save(this.user);

        assertTrue(userService.isUserExist(user));
    }

}
